리눅스 컴파일 테스트중

+++ solution_fidouaf_sdk_koscom_intergrateauth/UafMtsSdk/httputill.c
@@ -106,8 +106,8 @@ void saimplehttps(char* url, char* data) {
 }
 
 
-int httpsPost(char *pemcert, char* url, char* data,char** outData, size_t *outDataLen) {
-	int retData = 1;
+size_t httpsPost(char *pemcert, char* url, char* data,char** outData, size_t *outDataLen) {
+	size_t retData = 1;
 	CURL *curl;
 	CURLcode res;
 	struct curl_slist *headers = NULL;
